Fix 500 message on login

+++ app/controllers/sessions_controller.rb
@@ -3,8 +3,12 @@ class SessionsController < ApplicationController
   def create
     omniauth = request.env["omniauth.auth"]
     member = Member.authenticate(omniauth)
-    cookies[:access_token] = member.access_token if member
-    redirect_to session.delete(:return_to) || root_url
+    if member
+        cookies[:access_token] = member.access_token
+        redirect_to session.delete(:return_to) || root_url
+    else
+        redirect_to root_url, :alert => t('flash.login_failed')
+    end
   end
 
   def setup
